Utilities: move UniquePrefixMatcher; cleanup cmdline.
